+++ Sources/Navigator/EPUB/Preferences/EPUBSettings.swift
@@ -20,7 +20,10 @@ public struct EPUBSettings: ConfigurableSettings, Sendable {
     public var fontSize: Double
     public var fontWeight: Double?
     public var hyphens: Bool?
-    public var imageFilter: ImageFilter?
+    public var blendImages: Bool?
+    public var darkenImages: Double?
+    public var invertImages: Double?
+    public var invertGaiji: Double?
     public var language: Language?
     public var letterSpacing: Double?
     public var ligatures: Bool?
@@ -57,7 +60,10 @@ public struct EPUBSettings: ConfigurableSettings, Sendable {
         fontSize: Double,
         fontWeight: Double?,
         hyphens: Bool?,
-        imageFilter: ImageFilter?,
+        blendImages: Bool?,
+        darkenImages: Double?,
+        invertImages: Double?,
+        invertGaiji: Double?,
         language: Language?,
         letterSpacing: Double?,
         ligatures: Bool?,
@@ -87,7 +93,10 @@ public struct EPUBSettings: ConfigurableSettings, Sendable {
         self.fontSize = fontSize
         self.fontWeight = fontWeight
         self.hyphens = hyphens
-        self.imageFilter = imageFilter
+        self.blendImages = blendImages
+        self.darkenImages = darkenImages
+        self.invertImages = invertImages
+        self.invertGaiji = invertGaiji
         self.language = language
         self.letterSpacing = letterSpacing
         self.ligatures = ligatures
@@ -148,67 +157,65 @@ public struct EPUBSettings: ConfigurableSettings, Sendable {
             scroll = true
         }
 
+        let columnCount = preferences.columnCount ?? defaults.columnCount ?? 0
+        let fit = preferences.fit ?? defaults.fit ?? .auto
+        let fontSize = preferences.fontSize ?? defaults.fontSize ?? 1.0
+        let fontWeight = preferences.fontWeight ?? defaults.fontWeight
+        let hyphens = preferences.hyphens ?? defaults.hyphens
+        let blendImages = preferences.blendImages ?? defaults.blendImages
+        let darkenImages = preferences.darkenImages ?? defaults.darkenImages
+        let invertImages = preferences.invertImages ?? defaults.invertImages
+        let invertGaiji = preferences.invertGaiji ?? defaults.invertGaiji
+        let letterSpacing = preferences.letterSpacing ?? defaults.letterSpacing
+        let lineLength = preferences.lineLength ?? defaults.lineLength ?? 1.0
+        let ligatures = preferences.ligatures ?? defaults.ligatures
+        let lineHeight = preferences.lineHeight ?? defaults.lineHeight
+        let noRuby = preferences.noRuby ?? defaults.noRuby ?? false
+        let offsetFirstPage = preferences.offsetFirstPage ?? defaults.offsetFirstPage
+        let pageMargins = preferences.pageMargins ?? defaults.pageMargins ?? 1.0
+        let paragraphIndent = preferences.paragraphIndent ?? defaults.paragraphIndent
+        let paragraphSpacing = preferences.paragraphSpacing ?? defaults.paragraphSpacing
+        let publisherStyles = preferences.publisherStyles ?? defaults.publisherStyles ?? true
+        let spread = preferences.spread ?? defaults.spread ?? .auto
+        let textAlign = preferences.textAlign ?? defaults.textAlign
+        let textNormalization = preferences.textNormalization ?? defaults.textNormalization ?? false
+        let theme = preferences.theme ?? .light
+        let typeScale = preferences.typeScale ?? defaults.typeScale
+        let wordSpacing = preferences.wordSpacing ?? defaults.wordSpacing
+
         self.init(
             backgroundColor: preferences.backgroundColor,
-            columnCount: preferences.columnCount
-                ?? defaults.columnCount
-                ?? 0,
-            fit: preferences.fit
-                ?? defaults.fit
-                ?? .auto,
+            columnCount: columnCount,
+            fit: fit,
             fontFamily: preferences.fontFamily,
-            fontSize: preferences.fontSize
-                ?? defaults.fontSize
-                ?? 1.0,
-            fontWeight: preferences.fontWeight
-                ?? defaults.fontWeight,
-            hyphens: preferences.hyphens
-                ?? defaults.hyphens,
-            imageFilter: preferences.imageFilter
-                ?? defaults.imageFilter,
+            fontSize: fontSize,
+            fontWeight: fontWeight,
+            hyphens: hyphens,
+            blendImages: blendImages,
+            darkenImages: darkenImages,
+            invertImages: invertImages,
+            invertGaiji: invertGaiji,
             language: language,
-            letterSpacing: preferences.letterSpacing
-                ?? defaults.letterSpacing,
-            lineLength: preferences.lineLength
-                ?? defaults.lineLength
-                ?? 1.0,
-            ligatures: preferences.ligatures
-                ?? defaults.ligatures,
-            lineHeight: preferences.lineHeight
-                ?? defaults.lineHeight,
-            noRuby: preferences.noRuby
-                ?? defaults.noRuby
-                ?? false,
-            offsetFirstPage: preferences.offsetFirstPage
-                ?? defaults.offsetFirstPage,
-            pageMargins: preferences.pageMargins
-                ?? defaults.pageMargins
-                ?? 1.0,
-            paragraphIndent: preferences.paragraphIndent
-                ?? defaults.paragraphIndent,
-            paragraphSpacing: preferences.paragraphSpacing
-                ?? defaults.paragraphSpacing,
-            publisherStyles: preferences.publisherStyles
-                ?? defaults.publisherStyles
-                ?? true,
+            letterSpacing: letterSpacing,
+            ligatures: ligatures,
+            lineLength: lineLength,
+            lineHeight: lineHeight,
+            noRuby: noRuby,
+            offsetFirstPage: offsetFirstPage,
+            pageMargins: pageMargins,
+            paragraphIndent: paragraphIndent,
+            paragraphSpacing: paragraphSpacing,
+            publisherStyles: publisherStyles,
             readingProgression: readingProgression,
             scroll: scroll,
-            spread: preferences.spread
-                ?? defaults.spread
-                ?? .auto,
-            textAlign: preferences.textAlign
-                ?? defaults.textAlign,
+            spread: spread,
+            textAlign: textAlign,
             textColor: preferences.textColor,
-            textNormalization: preferences.textNormalization
-                ?? defaults.textNormalization
-                ?? false,
-            theme: preferences.theme
-                ?? .light,
-            typeScale: preferences.typeScale
-                ?? defaults.typeScale,
+            textNormalization: textNormalization,
+            theme: theme,
+            typeScale: typeScale,
             verticalText: verticalText,
-            wordSpacing: preferences.wordSpacing
-                ?? defaults.wordSpacing
+            wordSpacing: wordSpacing
         )
     }
 }
@@ -225,7 +232,10 @@ public struct EPUBDefaults: Sendable {
     public var fontSize: Double?
     public var fontWeight: Double?
     public var hyphens: Bool?
-    public var imageFilter: ImageFilter?
+    public var blendImages: Bool?
+    public var darkenImages: Double?
+    public var invertImages: Double?
+    public var invertGaiji: Double?
     public var language: Language?
     public var letterSpacing: Double?
     public var ligatures: Bool?
@@ -251,7 +261,10 @@ public struct EPUBDefaults: Sendable {
         fontSize: Double? = nil,
         fontWeight: Double? = nil,
         hyphens: Bool? = nil,
-        imageFilter: ImageFilter? = nil,
+        blendImages: Bool? = nil,
+        darkenImages: Double? = nil,
+        invertImages: Double? = nil,
+        invertGaiji: Double? = nil,
         language: Language? = nil,
         letterSpacing: Double? = nil,
         ligatures: Bool? = nil,
@@ -276,7 +289,10 @@ public struct EPUBDefaults: Sendable {
         self.fontSize = fontSize
         self.fontWeight = fontWeight
         self.hyphens = hyphens
-        self.imageFilter = imageFilter
+        self.blendImages = blendImages
+        self.darkenImages = darkenImages
+        self.invertImages = invertImages
+        self.invertGaiji = invertGaiji
         self.language = language
         self.letterSpacing = letterSpacing
         self.ligatures = ligatures