+++ Sources/Navigator/Preferences/Types.swift
@@ -92,12 +92,6 @@ public enum Theme: String, Codable, Hashable, Sendable {
     private static let sepiaBackgroundColor = Color(hex: "#faf4e8")!
 }
 
-/// Filter used to render images in a reflowable document.
-public enum ImageFilter: String, Codable, Hashable, Sendable {
-    case darken
-    case invert
-}
-
 /// Text alignment in a reflowable document.
 public enum TextAlignment: String, Codable, Hashable, Sendable {
     /// Align the text in the center of the page.