+++ Tests/NavigatorTests/EPUB/CSS/CSSUserPropertiesTests.swift
@@ -17,8 +17,10 @@ struct CSSUserPropertiesTests {
                 "--USER__pageMargins": nil,
                 "--USER__lineLength": nil,
                 "--USER__appearance": nil,
+                "--USER__blendImages": nil,
                 "--USER__darkenImages": nil,
                 "--USER__invertImages": nil,
+                "--USER__invertGaiji": nil,
                 "--USER__textColor": nil,
                 "--USER__backgroundColor": nil,
                 "--USER__fontOverride": nil,
@@ -48,8 +50,10 @@ struct CSSUserPropertiesTests {
                 pageMargins: 1.2,
                 lineLength: CSSPercentLength(42),
                 appearance: .night,
-                darkenImages: true,
-                invertImages: true,
+                blendImages: true,
+                darkenImages: 0.8,
+                invertImages: 0.9,
+                invertGaiji: 0.95,
                 textColor: CSSHexColor("#FF0000"),
                 backgroundColor: CSSHexColor("#00FF00"),
                 fontOverride: true,
@@ -71,10 +75,12 @@ struct CSSUserPropertiesTests {
                 "--USER__view": "readium-scroll-on",
                 "--USER__colCount": "0",
                 "--USER__pageMargins": "1.20000",
-                "--USER__lineLength": "4200%",
+                "--USER__lineLength": "4200.00000%",
                 "--USER__appearance": "readium-night-on",
-                "--USER__darkenImages": "readium-darken-on",
-                "--USER__invertImages": "readium-invert-on",
+                "--USER__blendImages": "readium-blend-on",
+                "--USER__darkenImages": "0.80000",
+                "--USER__invertImages": "0.90000",
+                "--USER__invertGaiji": "0.95000",
                 "--USER__textColor": "#FF0000",
                 "--USER__backgroundColor": "#00FF00",
                 "--USER__fontOverride": "readium-font-on",
@@ -134,8 +140,10 @@ struct CSSUserPropertiesTests {
                 pageMargins: 1.2,
                 lineLength: CSSPercentLength(42),
                 appearance: .night,
-                darkenImages: true,
-                invertImages: true,
+                blendImages: true,
+                darkenImages: 0.8,
+                invertImages: 0.9,
+                invertGaiji: 0.95,
                 textColor: CSSHexColor("#FF0000"),
                 backgroundColor: CSSHexColor("#00FF00"),
                 fontOverride: true,
@@ -158,16 +166,19 @@ struct CSSUserPropertiesTests {
             --USER__advancedSettings: readium-advanced-on !important;
             --USER__appearance: readium-night-on !important;
             --USER__backgroundColor: #00FF00 !important;
+            --USER__blendImages: readium-blend-on !important;
             --USER__bodyHyphens: auto !important;
             --USER__colCount: 0 !important;
-            --USER__darkenImages: readium-darken-on !important;
+            --USER__darkenImages: 0.80000 !important;
             --USER__fontFamily: "Times New", "Comic Sans" !important;
             --USER__fontOverride: readium-font-on !important;
             --USER__fontSize: 2.30000vmax !important;
-            --USER__invertImages: readium-invert-on !important;
+            --USER__invertGaiji: 0.95000 !important;
+            --USER__invertImages: 0.90000 !important;
             --USER__letterSpacing: 8.90000rem !important;
             --USER__ligatures: common-ligatures !important;
             --USER__lineHeight: 4.50000pt !important;
+            --USER__lineLength: 4200.00000% !important;
             --USER__noRuby: readium-noRuby-on !important;
             --USER__pageMargins: 1.20000 !important;
             --USER__paraIndent: 6.70000rem !important;